migrated segment_srv to ts
